Move IdGraphIndexCompute to its own file
